% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{Compute Enrollment Counts}
\usage{
compute(x)
}
\arguments{
\item{x}{A \code{selecta} object.}
}
\value{
A list with \code{nodes}, \code{edges}, and \code{phases} data.tables.
}
\description{
Compute Enrollment Counts
}
\keyword{internal}
