% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rctselect3}
\alias{rctselect3}
\title{Simulated Three-Arm Randomized Trial}
\format{
A \code{data.table} with 3,600 rows. See \code{\link{rctselect2}}
  for column descriptions.
}
\usage{
rctselect3
}
\description{
A synthetic dataset of 3,600 patients in a three-arm randomized
controlled trial. Structure matches \code{\link{rctselect2}} with an
additional treatment arm.
}
\examples{
data(rctselect3)
table(rctselect3$treatment)
}
\keyword{datasets}
