% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rctselect6}
\alias{rctselect6}
\title{Simulated Six-Arm Dose-Finding Trial}
\format{
A \code{data.table} with 7,200 rows. See \code{\link{rctselect2}}
  for column descriptions.
}
\usage{
rctselect6
}
\description{
A synthetic dataset of 7,200 patients in a six-arm dose-finding
trial. Structure matches \code{\link{rctselect2}} with six treatment
arms. Intended for demonstrating wide multi-arm layouts using
\code{\link{stratify}}.
}
\examples{
data(rctselect6)
table(rctselect6$treatment)
}
\keyword{datasets}
