% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selectaex3}
\alias{selectaex3}
\title{Simulated Three-Arm Randomized Trial}
\format{
A \code{data.table} with 2,400 rows. See \code{\link{selectaex2}}
  for column descriptions.
}
\usage{
selectaex3
}
\description{
A synthetic dataset of 2,400 patients in a three-arm randomized
controlled trial. Structure matches \code{\link{selectaex2}} with an
additional treatment arm.
}
\examples{
data(selectaex3)
str(selectaex3)
table(selectaex3$treatment)
}
\keyword{datasets}
