% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selectaex6}
\alias{selectaex6}
\title{Simulated Six-Arm Dose-Finding Trial}
\format{
A \code{data.table} with 3,600 rows. See \code{\link{selectaex2}}
  for column descriptions.
}
\usage{
selectaex6
}
\description{
A synthetic dataset of 3,600 patients in a six-arm dose-finding
trial. Structure matches \code{\link{selectaex2}} with six treatment
arms.
}
\examples{
data(selectaex6)
str(selectaex6)
table(selectaex6$treatment)
}
\keyword{datasets}
