% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_nodes}
\alias{layout_nodes}
\title{Layout Nodes for Grid Rendering}
\usage{
layout_nodes(graph)
}
\arguments{
\item{graph}{List from \code{compute()}.}
}
\value{
The graph with \code{x} and \code{row} columns on \code{nodes}.
}
\description{
Layout Nodes for Grid Rendering
}
\keyword{internal}
