% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.selecta}
\alias{print.selecta}
\title{Print an Enrollment Flow Summary}
\usage{
\method{print}{selecta}(x, ...)
}
\arguments{
\item{x}{A \code{selecta} object.}

\item{...}{Ignored.}
}
\description{
Print an Enrollment Flow Summary
}
