% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.selecta}
\alias{summary.selecta}
\title{Summarize an Enrollment Flow}
\usage{
\method{summary}{selecta}(object, ...)
}
\arguments{
\item{object}{A \code{selecta} object.}

\item{...}{Ignored.}
}
\value{
A \code{data.table} with columns: \code{phase}, \code{role},
  \code{arm}, \code{text}, \code{n}.
}
\description{
Computes all counts and returns a \code{data.table} summarizing each node.
}
