% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{to_dot}
\alias{to_dot}
\title{Convert Graph to Graphviz DOT String}
\usage{
to_dot(graph)
}
\arguments{
\item{graph}{A computed and laid-out graph.}
}
\value{
A character string in DOT format.
}
\description{
Convert Graph to Graphviz DOT String
}
\keyword{internal}
