% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.R
\name{to_dot}
\alias{to_dot}
\title{Convert Graph to Graphviz DOT String}
\usage{
to_dot(graph)
}
\arguments{
\item{graph}{A laid-out graph from \code{layout_nodes()}.}
}
\value{
A character string containing the DOT specification.
}
\description{
Convert Graph to Graphviz DOT String
}
\keyword{internal}
